requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
transformers
requests
python-dotenv
torch
